pyproject.toml
[project]
name = "gaboon"
version = "0.1"
description = "Pythonic smart contract development framework for Vyper."
dependencies = [
    "pytest>=8.2.2",
    "vyper>=0.4.0",
    "titanoboa @ git+https://github.com/vyperlang/titanoboa@vyper-0.4.0",
    "tomli-w>=1.0.0",
]
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Cyfrin" }]
maintainers = [{ name = "Cyfrin" }, { name = "Patrick Collins" }]

[project.scripts]
"gab" = "gaboon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "ruff>=0.5.2",
    "gaboon>=0.1",
    "mypy>=1.11.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["gaboon"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that make an actual RPC call (deselect with '-m \"not integration\"')",
]
addopts = "-m \"not integration\""

[tool.rye.scripts]
typecheck = "mypy ."

[tool.mypy]
disable_error_code = 'type-arg, attr-defined'
ignore_missing_imports = true
exclude = [
    'tests/',            # Excludes all files in the tests directory
    '(^|/)test_.*\.py$', # Excludes files that start with test_
]
